refactor: make dto directory for user and move files
