+++ springboot/src/main/kotlin/com/goliath/emojihub/springboot/dto/user/LoginRequest.kt
@@ -1,4 +1,4 @@
-package com.goliath.emojihub.springboot.dto
+package com.goliath.emojihub.springboot.dto.user
 
 data class LoginRequest (
     var username: String,
