+++ springboot/src/main/kotlin/com/goliath/emojihub/springboot/dto/user/SignUpRequest.kt
@@ -1,4 +1,4 @@
-package com.goliath.emojihub.springboot.dto
+package com.goliath.emojihub.springboot.dto.user
 
 data class SignUpRequest (
     var email: String,
